archival: wrap replication call in with_timeout_abortable
During the cluster shutdown it is possible that the raft replicate call
that persists archival metadata will hang for the entire timeout length.
To avoid that we wrap it with with_timeout_abortable so that it exits
quickly (when the abort source fires). This also prevents test failures
because timeout generates an error log message.